// src/types/list-types.ts
export enum Operations {
  Default = "default",
  AddToHead = "addToHead",
  AddToTail = "addToTail",
  RemoveFromHead = "removeFromHead",
  RemoveFromTail = "removeFromTail",
  InsertAtIndex = "insertAtIndex",
  RemoveAtIndex = "removeAtIndex",
}

export enum Positions {
  Top = "Top",
  Bottom = "Bottom",
}